add build_contacting_nn method
In some segmentation images, the radius parameter of the contacting_area
method has to be > 1 or 2 otherwise there are too many lonely nodes.
But then some nodes are linked *through* nodes in between.
Solution: this method first runs the are contact method with a small
distance parameter, then it runs the nearest neighbors method for the
lonely nodes, and adds the resulting edges to the previous edge list.